Remove plugin syntax and language overrides
These settings are better handled in a vimrc allowing pigment to only
establish color regions and a UI based off the color profile set in
Apple's Terminal preferences.

Without any overrides in your vimrc, the filetype's syntax is going to
be the point of truth. Sorry, you're colors are going to look a little
weird :).